requirements.txt
Django==2.2.3
ONEmSDK==0.3.6
PyJWT==1.7.1
django-heroku==0.3.1
gunicorn==19.9.0
requests==2.22.0
